+++ src/tooling/piral-cli/src/common/clients/lerna.ts
@@ -1,6 +1,6 @@
 import { resolve } from 'path';
 import { log } from '../log';
-import { findFile } from '../io';
+import { findFile, readJson, writeJson } from '../io';
 import { runCommand } from '../scripts';
 import { MemoryStream } from '../MemoryStream';
 
@@ -37,10 +37,34 @@ export async function installDependencies(target = '.', ...flags: Array<string>)
   return ms.value;
 }
 
+export async function uninstallPackage(packageRef: string, target = '.', ...flags: Array<string>) {
+  const ms = new MemoryStream();
+  // at the moment not supported by Lerna directly
+  // workaround is to manipulate the package.json directly and bootstrap again...
+  const packageData = await readJson(target, 'package.json');
+
+  if (packageData.dependencies && packageRef in packageData.dependencies) {
+    packageData.dependencies[packageRef] = undefined;
+  }
+
+  if (packageData.devDependencies && packageRef in packageData.devDependencies) {
+    packageData.devDependencies[packageRef] = undefined;
+  }
+
+  if (packageData.peerDependencies && packageRef in packageData.peerDependencies) {
+    packageData.peerDependencies[packageRef] = undefined;
+  }
+
+  await writeJson(target, 'package.json', packageData, true);
+  await runLernaProcess(['bootstrap'], target, ms);
+  log('generalDebug_0003', `Lerna bootstrap (after remove) package result: ${ms.value}`);
+  return ms.value;
+}
+
 export async function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runLernaProcess(['add', packageRef, ...convert(flags)], target, ms);
-  log('generalDebug_0003', `Lerna install package result: ${ms.value}`);
+  log('generalDebug_0003', `Lerna add package result: ${ms.value}`);
   return ms.value;
 }