+++ src/tooling/piral-cli/src/common/clients/rush.ts
@@ -40,10 +40,17 @@ export async function installDependencies(target = '.', ...flags: Array<string>)
   return ms.value;
 }
 
+export async function uninstallPackage(packageRef: string, target = '.', ...flags: Array<string>) {
+  const ms = new MemoryStream();
+  await runRushProcess(['remove', packageRef, ...convert(flags)], target, ms);
+  log('generalDebug_0003', `Rush remove package result: ${ms.value}`);
+  return ms.value;
+}
+
 export async function installPackage(packageRef: string, target = '.', ...flags: Array<string>) {
   const ms = new MemoryStream();
   await runRushProcess(['add', '--package', packageRef, ...convert(flags)], target, ms);
-  log('generalDebug_0003', `Rush install package result: ${ms.value}`);
+  log('generalDebug_0003', `Rush add package result: ${ms.value}`);
   return ms.value;
 }