+++ src/tooling/piral-cli/src/common/injectors.ts
@@ -2,16 +2,14 @@ import { resolve } from 'path';
 import { liveIcon, settingsIcon } from './emoji';
 import { logInfo, log, logReset } from './log';
 import { chalk } from '../external';
-import { Bundler } from '../types';
 
-export function notifyServerOnline(bundlers: Array<Bundler>, path: string, api: string | false) {
+export function notifyServerOnline(path: string, api: string | false) {
   return (svc: any) => {
     log('generalDebug_0003', `The kras server for debugging is online!`);
     const address = `${svc.protocol}://localhost:${chalk.green(svc.port)}`;
     logInfo(`${liveIcon} Running at ${chalk.bold(address + path)}`);
     logInfo(`${settingsIcon} Manage via ${chalk.bold(address + api)}`);
     logReset();
-    bundlers.forEach((bundler) => bundler.start());
   };
 }