+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 ## 0.15.1 (tbd)
 
 - Fixed update of `piral-extension` web component inside foreign components
+- Improved multi Piral instance debugging
 - Improved documentation
 - Added `pilet add-piral-instance` command to `piral-cli` (#543)
-- Added `pilet remove-piral-instance` command `piral-cli` (#543)
+- Added `pilet remove-piral-instance` command to `piral-cli` (#543)
 
 ## 0.15.0 (November 17, 2022)